% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PackageFunctions.R
\name{bedpe_to_rectangles_byChromosome}
\alias{bedpe_to_rectangles_byChromosome}
\title{Function to show you the whole chromosome, one window at a time.
It assumes that you started from a bedpe format file,
ran divide_bedpe_in_windows.sh to divide the file into genomic windows
window-level visualization, and are now ready to annotate each window.}
\usage{
bedpe_to_rectangles_byChromosome(bedpefile, chromo, mini, maxi, w, outdir, pref)
}
\arguments{
\item{bedpefile}{A template for the bedpe files that will be read.
Substitute WINDOW for the window size, and MINI for the beginning of the current window.
As example is: /srv/scratch/oursu/3Dgenome/results/processed_data/HiC/counts/intra/GM12878_combined/5kb/chr21/windowData/wWINDOW/chr21_5kb.RAWobserved.norm_SQRTVC.obsOverExp_SQRTVCexpected.wMINI.gz}

\item{chromo}{Name of the chromosome. This is for making the output file.}

\item{mini}{Starting position of the genomic windows to show.}

\item{maxi}{Ending position of the genomic windows to show}

\item{w}{Size of windows that will be shown to you.}

\item{outdir}{The name of the output directory where the output data will be stored}

\item{pref}{The prefix of the files that will be created. Useful for when the same method
will be run for finding different types of elements, e.g. TADs, loops, etc., then
you can just set the prefix to "TADs", "Loops", etc.}
}
\description{
Function to show you the whole chromosome, one window at a time.
It assumes that you started from a bedpe format file,
ran divide_bedpe_in_windows.sh to divide the file into genomic windows
window-level visualization, and are now ready to annotate each window.
}
\examples{
bedpe_to_rectangles_byChromosome('/srv/scratch/oursu/3Dgenome/results/processed_data/HiC/counts/intra/GM12878_combined/5kb/chr21/windowData/wWINDOW/chr21_5kb.RAWobserved.norm_SQRTVC.obsOverExp_SQRTVCexpected.wMINI.gz','chr21',30000000,39000000,1000000,outdir,outpref)
}

