% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_rectangularized_coords}
\alias{get_rectangularized_coords}
\title{Convert click coordinates to rectangles.
This function assumes that click coordinates are given in pairs
The first click is the lower left, and the second click is the top right}
\usage{
get_rectangularized_coords(coords, chromo, mini, maxi)
}
\arguments{
\item{coords}{The coordinates object received from pairs of clicks identifying rectangles.}

\item{chromo}{Current chromosome for the rectangles (for making a bedpe file of rectangles).}

\item{mini}{Minimum coordinate in the figure (important for the conversion from pixels to genomic coordinates)}

\item{maxi}{Maximum coordinate in the figure}
}
\value{
A bedpe data frame with rectangles (first bedpe entry is the y axis, second bedpe entry is the x axis)
}
\description{
Convert click coordinates to rectangles.
This function assumes that click coordinates are given in pairs
The first click is the lower left, and the second click is the top right
}

